+++ blackbox.php
@@ -12,34 +12,6 @@ use Innmind\BlackBox\{
 };
 use function Innmind\BlackBox\Runner\test;
 
-// This test has to be done here because other tests use global functions
-$result = Application::new([])
-    ->disableGlobalFunctions()
-    ->tryToProve(function($prove) {
-        yield $prove->test(
-            'Global functions can be disabled',
-            static fn($assert) => $assert->true(
-                Application::new([])
-                    ->disableGlobalFunctions()
-                    ->displayOutputVia(Collect::new())
-                    ->displayErrorVia(Collect::new())
-                    ->tryToProve(function() {
-                        yield test(
-                            'Test function is not declared',
-                            static fn($assert) => $assert->false(
-                                \function_exists('test'),
-                            ),
-                        );
-                    })
-                    ->successful(),
-            ),
-        );
-    });
-
-if (!$result->successful()) {
-    $result->exit();
-}
-
 Application::new($argv)
     ->map(static fn($app) => match (\getenv('BLACKBOX_ENV')) {
         'coverage' => $app